Split HttpCodesApi example up
